requirements.txt
anthropic>=0.25.0
pyyaml>=6.0
pytest>=8.0
